Added project view
This is a new screen, which dispalys all images from the project. In the future, it'll allow editing, sorting and much more.

For support to be added, some changed had to be made to the ProjectControl, pressing on a project in the project manager will now open a new screen with the project.